Fetched Constellations from SNACC for all ids
First fetched a single constellation and then, with the help of Nabil, fetched constellation for each ids. The ids are separated and kept in a json file under data. Moreover, a Gatsby node is created for each of the constellation and made it available to graphql for querying. This is achieved to complete the data parsing so they can be served to components.