Progress on Transcript View
Done:
Created buttons that paginates bidirectionally to the next image of the
currently rendered journals.
Created buttons that "paginates" to the next journal file bidirectionally

Need for the current topic:
When the image is being paginated, paginate the corresponding page break;
based on the pid, or n (pb index), or both.