Extend highlighting of Hunt to arbitrary nodes
The network diagram previously accepted a single node ID in its centralFigure prop. This central figure received special treatment in a number of ways. The recoding extends this, allowing an arbitrary number of central figures to be supplied. This commit also adds Joshua Evans to the list of central figures currently being used.